2. Responsi/app/requirements.txt
Flask

mysql-connector
